+++ packages/mongo-nav/src/helpers/OnPremUserMenu.tsx
@@ -90,6 +90,7 @@ export default function OnPremUserMenu({
           href={urls.onPrem.profile}
           active={activeNav === ActiveNavElement.UserMenuOnPremProfile}
           data-testid="om-user-menuitem-profile"
+          onClick={() => setOpen(false)}
         >
           Profile
         </MenuItem>
@@ -99,6 +100,7 @@ export default function OnPremUserMenu({
             href={urls.onPrem.mfa}
             active={activeNav === ActiveNavElement.UserMenuOnPremTwoFactorAuth}
             data-testid="om-user-menuitem-mfa"
+            onClick={() => setOpen(false)}
           >
             Two-factor Authentication
           </MenuItem>
@@ -108,6 +110,7 @@ export default function OnPremUserMenu({
           href={urls.onPrem.personalization}
           active={activeNav === ActiveNavElement.UserMenuOnPremPersonalization}
           data-testid="om-user-menuitem-personalization"
+          onClick={() => setOpen(false)}
         >
           Personalization
         </MenuItem>
@@ -116,6 +119,7 @@ export default function OnPremUserMenu({
           href={urls.onPrem.invitations}
           active={activeNav === ActiveNavElement.UserMenuOnPremInvitations}
           data-testid="om-user-menuitem-invitations"
+          onClick={() => setOpen(false)}
         >
           Invitations
         </MenuItem>
@@ -124,6 +128,7 @@ export default function OnPremUserMenu({
           href={urls.onPrem.organizations}
           active={activeNav === ActiveNavElement.UserMenuOnPremOrganizations}
           data-testid="om-user-menuitem-organizations"
+          onClick={() => setOpen(false)}
         >
           Organizations
         </MenuItem>
@@ -131,12 +136,13 @@ export default function OnPremUserMenu({
         <MenuItem
           href={urls.onPrem.featureRequest}
           data-testid="om-user-menuitem-feature-request"
+          onClick={() => setOpen(false)}
         >
           Feature Request
         </MenuItem>
 
         <MenuItem
-          onClick={onElementClick(NavElement.Logout)}
+          onClick={onElementClick(NavElement.Logout, () => setOpen(false))}
           data-testid="om-user-menuitem-sign-out"
         >
           Sign Out