+++ packages/mongo-nav/src/project-nav/ProjectNav.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState } from 'react';
 import Tooltip from '@leafygreen-ui/tooltip';
 import IconButton from '@leafygreen-ui/icon-button';
 import Icon from '@leafygreen-ui/icon';
@@ -9,8 +9,8 @@ import { uiColors } from '@leafygreen-ui/palette';
 import { useUsingKeyboardContext } from '@leafygreen-ui/leafygreen-provider';
 import { ProjectSelect } from '../mongo-select';
 import { facepaint, breakpoints } from '../breakpoints';
-import { useViewportSize } from '@leafygreen-ui/hooks';
-import { iconLoadingStyle, textLoadingStyle } from '../styles';
+import { useViewportSize, usePoller } from '@leafygreen-ui/hooks';
+import { iconLoadingStyle, textLoadingStyle, anchorOverrides } from '../styles';
 import { useOnElementClick } from '../on-element-click-provider';
 import {
   ProjectInterface,
@@ -21,9 +21,16 @@ import {
   PlanType,
   ActiveNavElement,
   NavElement,
+  Mode,
   MongoNavInterface,
+  ProjectStatus,
 } from '../types';
-import { AtlasIcon, RealmIcon, ChartsIcon } from '../helpers/Icons';
+import {
+  AtlasIcon,
+  RealmIcon,
+  ChartsIcon,
+  ProjectStatusBadge,
+} from '../helpers';
 
 const {
   ProjectNavProjectDropdown,
@@ -42,6 +49,16 @@ export const projectNavHeight = 45;
 
 const productIconProp = createDataProp('charts-data-prop');
 
+const projectNavAnchorOverrides = css`
+  a {
+    &:visited,
+    &:active,
+    &:link {
+      color: ${uiColors.gray.dark2};
+    }
+  }
+`;
+
 const navContainerStyle = css`
   display: flex;
   align-items: center;
@@ -59,7 +76,7 @@ const navContainerStyle = css`
 const menuIconButtonStyle = css`
   z-index: 1;
   ${facepaint({
-    marginRight: ['20px', '14px', '20px'],
+    marginRight: ['16px', '14px', '16px'],
   })}
 `;
 
@@ -223,7 +240,7 @@ const secondTabName = displayProductName();
 
 type ProjectNavProps = Pick<
   MongoNavInterface,
-  'activeProduct' | 'activeNav' | 'onProjectChange'
+  'activeProduct' | 'activeNav' | 'onProjectChange' | 'admin' | 'mode'
 > & {
   current?: CurrentProjectInterface;
   data?: Array<ProjectInterface>;
@@ -237,6 +254,7 @@ type ProjectNavProps = Pick<
 };
 
 export default function ProjectNav({
+  admin,
   current,
   data,
   constructProjectURL,
@@ -245,9 +263,11 @@ export default function ProjectNav({
   activeNav,
   onProjectChange,
   hosts,
-  alerts = 0,
+  mode,
 }: ProjectNavProps) {
-  const [open, setOpen] = React.useState(false);
+  const [open, setOpen] = useState(false);
+  const [alerts, setAlerts] = useState(current?.alertsOpen ?? 0);
+
   const { usingKeyboard: showFocus } = useUsingKeyboardContext();
   const { width: viewportWidth } = useViewportSize();
   const onElementClick = useOnElementClick();
@@ -256,6 +276,32 @@ export default function ProjectNav({
   const isCloudManager = current?.planType === PlanType.Cloud;
   const isLoading = !!current;
 
+  const currentProjectId = current?.projectId;
+
+  function fetchAlertsCount() {
+    return fetch(
+      `${hosts.cloud}/user/shared/alerts/project/${currentProjectId}`,
+      {
+        credentials: 'include',
+        mode: 'cors',
+        method: 'GET',
+      },
+    )
+      .then(response => response.json())
+      .then(result => setAlerts(result.length))
+      .catch(console.error);
+  }
+
+  usePoller(fetchAlertsCount, {
+    interval:
+      activeProduct === Product.Cloud
+        ? // 2 minutes for Atlas
+          120e3
+        : // 10 minutes for non-Atlas
+          600e3,
+    enabled: mode === Mode.Production && currentProjectId != null,
+  });
+
   const sharedTooltipProps = {
     variant: 'dark',
     usePortal: false,
@@ -263,7 +309,7 @@ export default function ProjectNav({
   } as const;
 
   const getProductClassName = (product: Product) =>
-    cx(productStyle, {
+    cx(anchorOverrides, projectNavAnchorOverrides, productStyle, {
       [productStates.active]: !!(activeProduct === product && current),
       [productStates.focus]: showFocus,
       [productStates.loading]: !current,
@@ -338,6 +384,13 @@ export default function ProjectNav({
           Integrations
         </MenuItem>
       </Menu>
+
+      {!isMobile &&
+        admin &&
+        current?.status &&
+        Object.values(ProjectStatus).includes(current?.status) && (
+          <ProjectStatusBadge currentStatus={current.status} />
+        )}
 
       <ul className={productListStyle}>
         <li role="none" className={productTabStyle}>