+++ packages/mongo-nav/src/types.ts
@@ -150,6 +150,17 @@ type OrgPaymentLabel = typeof OrgPaymentLabel[keyof typeof OrgPaymentLabel];
 
 export { OrgPaymentLabel };
 
+const ProjectStatus = {
+  Active: 'ACTIVE',
+  Dead: 'DEAD',
+  Closing: 'CLOSING',
+  Closed: 'CLOSED',
+};
+
+type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];
+
+export { ProjectStatus };
+
 export interface AccountInterface {
   email: string;
   firstName: string;
@@ -168,6 +179,7 @@ export interface ProjectInterface {
 export interface CurrentProjectInterface extends ProjectInterface {
   alertsOpen: number;
   chartsActivated: boolean;
+  status?: ProjectStatus;
 }
 
 export interface OrganizationInterface {
@@ -355,9 +367,4 @@ export interface MongoNavInterface {
    * Determines whether or not the component will fetch data from cloud
    */
   loadData?: boolean;
-
-  /**
-   * Overwrite number of alerts received from cloud endpoint
-   */
-  alertsCount?: number;
 }