+++ typings/jest.d.ts
@@ -1,0 +1,5 @@
+declare namespace jest {
+  interface Matchers<R, T> {
+    toBePx: (expected: number) => R;
+  }
+}